add mensa api request
