+++ .gitignore
@@ -87,3 +87,4 @@ assets/.DS_Store
 
 credentials.yml
 
+spluseinscache
