+++ server/controllers/splus.ts
@@ -8,11 +8,13 @@ import { SplusApi } from '../lib/SplusApi';
 import { RichLecture } from '../model/RichLecture';
 
 const SCHEDULE_CACHE_SECONDS = 600;
+const MENSA_CACHE_SECONDS = 3600;
 
 // default must be in /tmp because the rest is RO on AWS Lambda
 const CACHE_PATH = process.env.CACHE_PATH || '/tmp/spluseins-cache';
 const CACHE_DISABLE = !!process.env.CACHE_DISABLE;
 
+const axios = require('axios');
 const router = express.Router();
 const cache = CACHE_DISABLE ?
   cacheManager.caching({ store: 'memory', max: 0 }) :
@@ -29,6 +31,7 @@ const cache = CACHE_DISABLE ?
  * Accept CORS preflight requests.
  */
 router.options('/:schedule/:week', cors());
+router.options('/mensa', cors());
 
 /**
  * Get all lectures for the given schedule and week.
@@ -44,12 +47,45 @@ router.get('/:schedule/:week', cors(), async (req, res, next) => {
 
   try {
     const data = await cache.wrap(key, async () => {
-      console.log(`cache miss for key ${key}`);
+      console.log(`timetable cache miss for key ${key}`);
       const lectures = await SplusApi.getData('#' + schedule, week);
       return lectures.map((ilecture) => new RichLecture(ilecture, week));
     }, { ttl: SCHEDULE_CACHE_SECONDS });
 
     res.set('Cache-Control', `public, max-age=${SCHEDULE_CACHE_SECONDS}`);
+    res.json(data);
+  } catch (error) {
+    next(error);
+  }
+});
+
+
+router.get('/mensa', cors(), async (req, res, next) => {
+  const key = (new Date()).toString();
+  let openDays = [];
+  let weekdays = [];
+  let result = [];
+
+  try {
+    const data = await cache.wrap(key, async () => {
+      console.log(`mensa cache miss for key ${key}`);
+
+      const response = await axios.get(`https://openmensa.org/api/v2/canteens/166/days`);
+      openDays = response.data;
+
+      for(let i=0; i<3; i++) {
+        weekdays.push(moment(openDays[i].date));
+      }
+  
+      await Promise.all(weekdays.map(async (day) => {
+        const response = await axios.get(`https://openmensa.org/api/v2/canteens/166/days/${day.format('YYYY-MM-DD')}/meals`);
+        result.push({id: day.month() + day.day(), date: day.format('YYYY-MM-DD'), data: {...response.data}});
+      }));
+
+      return result.sort((a,b) => a.id - b.id);
+    }, { ttl: MENSA_CACHE_SECONDS });
+
+    res.set('Cache-Control', `public, max-age=${MENSA_CACHE_SECONDS}`);
     res.json(data);
   } catch (error) {
     next(error);
