+++ store/mensa.js
@@ -12,31 +12,15 @@ export const mutations = {
 
 export const actions = {
   async loadWeek({ state, commit }) {
-    let openDays = [];
-    let weekdays = [];
     let result = [];
 
     try {
-      const response = await this.$axios.get(`https://openmensa.org/api/v2/canteens/166/days`);
-      openDays = response.data;
+      const response = await this.$axios.get(`/api/splus/mensa`);
+      result = response.data;
     } catch (error) {
+        commit('setError', 'API-Verbindung fehlgeschlagen');
         console.error('error during Mensa API call', error.message);
     }
-
-    for(let i=0; i<3; i++) {
-        weekdays.push(moment(openDays[i].date));
-    }
-
-    await Promise.all(weekdays.map(async (day) => {
-        try {
-            const response = await this.$axios.get(`https://openmensa.org/api/v2/canteens/166/days/${day.format('YYYY-MM-DD')}/meals`);
-            result.push({date: day, data: {...response.data}});
-        } catch (error) {
-            console.error('error during Mensa API call', error.message);
-        }
-    }));
-
-    result = result.sort((a,b) => a.date.day() > b.date.day());
 
     commit('setWeekPlan', result);
   }
